+++ src/pages/folder/edit/index.tsx
@@ -1,25 +1,21 @@
 import Header from '@/components/common/Header';
 import { NextPage } from 'next';
 import styled from 'styled-components';
-import { FOLDERLIST } from '@/constants/folderList';
-import { colorDecoder } from '@/utils/folder';
 import { useState } from 'react';
 import { Folder } from '@/types/folder';
 import EditModal from '@/components/folder/EditModal';
-import { useRouter } from 'next/router';
 import SearchSection from '@/components/common/SearchSection';
+import { useGetFolder } from '@/hooks/api/useFolder';
+import RenderFolderList from '@/components/folder/RenderFolderList';
 
 interface Props {}
 
 const FolderEdit: NextPage<Props> = ({}) => {
-  const router = useRouter();
   const [searchInput, setSearchInput] = useState('');
-  const [newFolderList, setNewFolderList] = useState<Folder[]>(FOLDERLIST);
+  const { data } = useGetFolder();
+  const [newFolderList, setNewFolderList] = useState<Folder[]>(data as Folder[]);
   const [searchedFolderList, setSearchedFolderList] = useState<Folder[]>([]);
-  const folderNameList = newFolderList.map((folder) => folder.folderName);
-  const [isEditingFolder, setIsEditingFolder] = useState('');
   const [targetFolder, setTargetFolder] = useState<Folder>(newFolderList[0]);
-  const [newFolderName, setNewFolderName] = useState('');
   const [isModalOn, setIsModalOn] = useState(false);
 
   const handleSearch = (value: string) => {
@@ -29,59 +25,9 @@ const FolderEdit: NextPage<Props> = ({}) => {
     );
   };
 
-  const handleBlur = () => {
-    setNewFolderName('');
-    setIsEditingFolder('');
-  };
-
-  const checkEnter = (key: string, folder: Folder) => {
-    if (key === 'Enter') {
-      if (folderNameList.includes(newFolderName)) {
-        alert('이미 같은 이름의 폴더가 존재합니다!');
-        handleBlur();
-        return;
-      }
-      const newList = newFolderList;
-      newList[newFolderList.indexOf(folder)] = {
-        ...folder,
-        folderName: newFolderName,
-      };
-      setNewFolderList(newList);
-      setNewFolderName('');
-      setIsEditingFolder('');
-    }
-  };
-
   const handleModalOn = (folder: Folder) => {
     setTargetFolder(folder);
     setIsModalOn(true);
-  };
-
-  const handleColorChange = () => {
-    router.push(
-      {
-        pathname: '/folder/edit-color',
-        query: {
-          folderId: targetFolder.folderId,
-          folderName: targetFolder.folderName,
-          folderColor: targetFolder.folderColor,
-          insightCount: targetFolder.insightCount,
-        },
-      },
-      '/folder/edit-color',
-    );
-  };
-
-  const handleDelete = () => {
-    router.push(
-      {
-        pathname: '/folder/delete',
-        query: {
-          folderId: targetFolder.folderId,
-        },
-      },
-      '/folder/delete',
-    );
   };
 
   const saveFolder = () => {};
@@ -100,74 +46,17 @@ const FolderEdit: NextPage<Props> = ({}) => {
           bottom={20}
         />
         <FolderSection>
-          {searchInput === '' &&
-            newFolderList.map((folder, idx) => (
-              <FolderRow key={idx}>
-                {colorDecoder(folder.folderColor, 'small')}
-                <div className="text-container">
-                  {isEditingFolder === folder.folderName ? (
-                    <Input
-                      type="text"
-                      value={newFolderName}
-                      onChange={(e) => setNewFolderName(e.target.value)}
-                      onKeyDownCapture={(e) => checkEnter(e.key, folder)}
-                      onBlur={handleBlur}
-                      placeholder={folder.folderName}
-                    />
-                  ) : (
-                    <span
-                      className="name text"
-                      onClick={() => setIsEditingFolder(folder.folderName)}
-                    >
-                      {folder.folderName}
-                    </span>
-                  )}
-                  <span
-                    className="edit text"
-                    onClick={() => handleModalOn(folder)}
-                  >
-                    편집
-                  </span>
-                </div>
-              </FolderRow>
-            ))}
-          {searchInput !== '' &&
-            searchedFolderList.map((folder, idx) => (
-              <FolderRow key={idx}>
-                {colorDecoder(folder.folderColor, 'small')}
-                <div className="text-container">
-                  {isEditingFolder === folder.folderName ? (
-                    <Input
-                      type="text"
-                      value={newFolderName}
-                      onChange={(e) => setNewFolderName(e.target.value)}
-                      onKeyDownCapture={(e) => checkEnter(e.key, folder)}
-                      onBlur={handleBlur}
-                      placeholder={folder.folderName}
-                    />
-                  ) : (
-                    <span
-                      className="name text"
-                      onClick={() => setIsEditingFolder(folder.folderName)}
-                    >
-                      {folder.folderName}
-                    </span>
-                  )}
-                  <span
-                    className="edit text"
-                    onClick={() => handleModalOn(folder)}
-                  >
-                    편집
-                  </span>
-                </div>
-              </FolderRow>
-            ))}
+          <RenderFolderList
+            folderList={searchInput === '' ? newFolderList : searchedFolderList}
+            handleModalOn={handleModalOn}
+            newFolderList={newFolderList}
+            setNewFolderList={setNewFolderList}
+          />
         </FolderSection>
         {isModalOn && (
           <EditModal
             type="edit"
-            onClick1={() => handleColorChange()}
-            onClick2={() => handleDelete()}
+            targetFolder={targetFolder}
             onClose={() => setIsModalOn(false)}
           />
         )}
@@ -216,44 +105,3 @@ const FolderSection = styled.div`
   gap: 17px;
 `;
 
-const FolderRow = styled.div`
-  display: flex;
-  margin: auto;
-  flex-direction: row;
-  width: calc(100% - 40px);
-  .text-container {
-    width: calc(100% - 100px - 40px);
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    margin-left: 28px;
-  }
-  .text {
-    font-family: Pretendard;
-    font-style: normal;
-    line-height: 140%; /* 25.2px */
-  }
-  .name {
-    color: var(--Neutral-500, #1f1f1f);
-
-    font-size: 18px;
-    font-weight: 700;
-  }
-  .edit {
-    color: var(--Neutral-300, #848484);
-    font-size: 16px;
-    font-weight: 600;
-  }
-`;
-
-const Input = styled.input`
-  border: none;
-  outline: none;
-
-  font-family: Pretendard;
-  font-size: 17px;
-  font-style: normal;
-  font-weight: 600;
-  line-height: 21px; /* 123.529% */
-  letter-spacing: -0.32px;
-`;
